requirements.txt
music21==7.1.0
pyyaml==6.0.2
streamlit